+++ android/src/main/kotlin/com/superwall/superwallkit_flutter/utils/OptionsMapper.kt
@@ -133,6 +133,10 @@ fun PPaywallOptions.toSdkPaywallOptions(): PaywallOptions {
             }
     }
 
+    this.overrideProductsByName?.let { hostOverrides ->
+        sdkPaywallOptions.overrideProductsByName = hostOverrides
+    }
+
     return sdkPaywallOptions
 }